Modify Get and Delete endpoints (catch error)

+++ src/products/products.controller.ts
@@ -23,11 +23,15 @@ export class ProductsController {
   }
   @Get('/:id')
   getById(@Param('id', new ParseUUIDPipe()) id: string) {
-    return this.productsService.getById(id);
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
     this.productsService.deleteById(id);
     return { success: true };
   }
